microservices/auth/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.16.2",
    "argon2-cffi>=25.1.0",
    "asyncpg==0.30.0",
    "celery>=5.5.3",
    "fastapi[standard]==0.115.14",
    "greenlet==3.2.3",
    "itsdangerous>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]==3.2.9",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.1.1",
    "python-jose>=3.5.0",
    "redis>=6.2.0",
    "sqlalchemy==2.0.41",
    "uvicorn[standard]>=0.35.0",
]
[tool.setuptools]
packages=["app"]
[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "ruff>=0.12.1",
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
python_files = "test_*.py"

#[tool.uv.scripts]
#lint = "ruff check"
#fmt = "ruff format"
#typecheck = "mypy"
#t = "pytest"
#dev = "fastapi dev app/main.py"

[tool.mypy]
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
no_implicit_optional = true
